+++ src/Core/BackgroundSchedulePoolTaskHolder.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <memory>
+#include <mutex>
 
 namespace DB
 {
@@ -32,6 +33,39 @@ public:
 
 private:
     BackgroundSchedulePoolTaskInfoPtr task_info;
+};
+
+/// RAII guard that pauses parts check and reactivates it on destruction.
+/// Safe to destroy from any thread.
+class PausableTask
+{
+public:
+    explicit PausableTask(BackgroundSchedulePoolTaskHolder task_);
+
+    PausableTask(const PausableTask &) = delete;
+    PausableTask & operator=(const PausableTask &) = delete;
+    PausableTask(PausableTask &&) = delete;
+    PausableTask & operator=(PausableTask &&) = delete;
+
+    BackgroundSchedulePoolTaskHolder & getTask();
+
+    void pause();
+    void resume();
+
+    struct PauseHolder
+    {
+        explicit PauseHolder(PausableTask & task_);
+        ~PauseHolder();
+
+        private:
+            PausableTask & task;
+    };
+
+    using PauseHolderPtr = std::unique_ptr<PauseHolder>;
+private:
+    std::mutex pause_mutex;
+    size_t pause_count = 0;
+    BackgroundSchedulePoolTaskHolder task;
 };
 
 }
